+++ apps/web/lib/tinybird/index.ts
@@ -1,5 +1,5 @@
 export * from "./client";
-export * from "./delete-events";
+export * from "./delete-customer-events";
 export * from "./get-click-event";
 export * from "./get-lead-event";
 export * from "./record-click";
